refactor commits
